+++ packages/apollo-env/src/index.ts
@@ -2,4 +2,3 @@ import "./polyfills";
 
 export * from "./typescript-utility-types";
 export * from "./fetch";
-export * from "./utils";
