--- packages/apollo-env/src/utils/predicates.ts
@@ -1,5 +1,0 @@
-export function isNotNullOrUndefined<T>(
-  value: T | null | undefined
-): value is T {
-  return value !== null && typeof value !== "undefined";
-}
