+++ packages/apollo-graphql/package.json
@@ -11,7 +11,6 @@
     "node": ">=6"
   },
   "dependencies": {
-    "apollo-env": "file:../apollo-env",
     "lodash.sortby": "^4.7.0"
   },
   "peerDependencies": {
