+++ packages/apollo-graphql/src/utilities/isNodeLike.ts
